backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
pytesseract
pillow
openai
anthropic
PyPDF2
